refactor(app): split diplomacy and civilian panels under ~700 lines
Extract diplomacy row model and buildDiplomacyRows into diplomacy_panel_rows.dart.
Move civilian panel helpers and unit row widgets into part civilian_units_panel_support.dart.

Refs #1878

Made-with: Cursor
