+++ app/lib/features/game/widgets/diplomacy_panel.dart
@@ -1,9 +1,8 @@
 // Diplomacy panel. SPEC/ui/diplomacy-panel.md.
 
-import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_ai/colonizethis_ai.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
-import 'package:colonizethis_logic/order_suggestion_api.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:flutter/material.dart';
 
@@ -14,210 +13,10 @@ import '../../../l10n/l10n.dart';
 import '../../../widgets/ct_nine_patch_button.dart';
 import '../../../widgets/ct_panel.dart';
 import 'diplomacy_order_helpers.dart';
+import 'diplomacy_panel_rows.dart';
 import 'fnv1a_hash_constants.dart';
 
-int? _outgoingSubsidyPerTurn(Game game, String payerId, String targetId) {
-  for (final s in game.subsidyStates) {
-    if (s.payerId == payerId && s.targetId == targetId) {
-      return s.amountPerTurn;
-    }
-  }
-  return null;
-}
-
-({int? grant, int? subsidy}) _pendingEconomicAmounts(
-  List<DiplomaticOrder> list,
-  String targetId,
-) {
-  int? grant;
-  int? subsidy;
-  for (final o in list) {
-    if (o.targetFactionId != targetId) continue;
-    if (o.type == DiplomaticOrderType.grantAid) {
-      grant = o.amount;
-    }
-    if (o.type == DiplomaticOrderType.setSubsidy) {
-      subsidy = o.amount;
-    }
-  }
-  return (grant: grant, subsidy: subsidy);
-}
-
-/// Faction type for display. SPEC/game/factions.md.
-enum FactionKind { greatPower, minor, tribe }
-
-/// One row of data for the diplomacy list.
-class DiplomacyRowData {
-  const DiplomacyRowData({
-    required this.factionId,
-    required this.displayName,
-    required this.kind,
-    required this.relation,
-    this.overture,
-    required this.actions,
-    this.powerScore,
-    this.playerPowerScore,
-    required this.pendingOrderTypes,
-    this.activeSubsidyPerTurn,
-    this.pendingGrantAmount,
-    this.pendingSubsidyAmount,
-  });
-
-  final String factionId;
-  final String displayName;
-  final FactionKind kind;
-  final DiplomacyRelation? relation;
-  final OvertureState? overture;
-  final List<DiplomaticOrder> actions;
-
-  /// Great Power power score (SPEC/game/diplomacy.md). Only set for GP rows.
-  final int? powerScore;
-
-  /// Human player's power score for comparison (red if GP score > this). Only set for GP rows.
-  final int? playerPowerScore;
-
-  /// Set of DiplomaticOrderType that are currently pending for this target faction.
-  final Set<DiplomaticOrderType> pendingOrderTypes;
-
-  /// Active £/turn subsidy from the human GP to this row's faction (`Game.subsidyStates`).
-  final int? activeSubsidyPerTurn;
-
-  /// Pending grant aid amount in current-turn orders (not yet resolved).
-  final int? pendingGrantAmount;
-
-  /// Pending set-subsidy amount per turn in current-turn orders.
-  final int? pendingSubsidyAmount;
-}
-
-/// Builds list of discovered factions and their available actions.
-/// Discovered = has a relation with the player. SPEC/ui/diplomacy-panel.md.
-List<DiplomacyRowData> buildDiplomacyRows(
-  Game game,
-  MapTopology topology,
-  String humanPlayerId,
-  Orders currentOrders,
-) {
-  const suggestionApi = DefaultOrderSuggestionAPI();
-  final view = buildPlayerView(game, topology, humanPlayerId);
-  final discoveredIds = view.diplomacyByOtherId.keys.toList();
-  final suggestions = suggestionApi.suggestDiplomaticOrders(
-    view,
-    game,
-    topology,
-    currentOrders,
-  );
-  final actionsByTarget = <String, List<DiplomaticOrder>>{};
-  for (final order in suggestions) {
-    actionsByTarget.putIfAbsent(order.targetFactionId, () => []).add(order);
-  }
-
-  final gpIds = <String>[];
-  final minorIds = <String>[];
-  final tribeIds = <String>[];
-  for (final id in discoveredIds) {
-    if (id == humanPlayerId) continue;
-    if (game.players.any((p) => p.id == id)) {
-      gpIds.add(id);
-    } else if (game.minorNations.any((m) => m.id == id)) {
-      minorIds.add(id);
-    } else if (game.tribes.any((t) => t.id == id)) {
-      tribeIds.add(id);
-    }
-  }
-
-  // GPs: sort by military power desc, then province count desc. SPEC/ui/diplomacy-panel.md.
-  gpIds.sort((a, b) {
-    final strA = aggregateMilitaryStrengthForPlayer(game, a);
-    final strB = aggregateMilitaryStrengthForPlayer(game, b);
-    final cmp = strB.compareTo(strA);
-    if (cmp != 0) return cmp;
-    final provA = provinceCountOwnedBy(game, a);
-    final provB = provinceCountOwnedBy(game, b);
-    return provB.compareTo(provA);
-  });
-
-  final pendingByTarget = <String, Set<DiplomaticOrderType>>{};
-  final pendingList =
-      currentOrders.diplomaticOrdersByPlayerId[humanPlayerId] ?? [];
-  for (final o in pendingList) {
-    pendingByTarget.putIfAbsent(o.targetFactionId, () => {}).add(o.type);
-  }
-
-  String displayNameFor(String id) {
-    final p = game.playerById(id);
-    if (p != null) return p.displayName;
-    for (final m in game.minorNations) {
-      if (m.id == id) return m.displayName ?? id;
-    }
-    for (final t in game.tribes) {
-      if (t.id == id) return t.displayName ?? id;
-    }
-    return id;
-  }
-
-  List<DiplomacyRowData> rows = [];
-  final playerPower = greatPowerPowerScore(game, humanPlayerId);
-  for (final id in gpIds) {
-    final econ = _pendingEconomicAmounts(pendingList, id);
-    rows.add(
-      DiplomacyRowData(
-        factionId: id,
-        displayName: displayNameFor(id),
-        kind: FactionKind.greatPower,
-        relation: view.diplomacyByOtherId[id],
-        overture: getOverture(game, humanPlayerId, id),
-        actions: actionsByTarget[id] ?? [],
-        powerScore: greatPowerPowerScore(game, id),
-        playerPowerScore: playerPower,
-        pendingOrderTypes: pendingByTarget[id] ?? {},
-        activeSubsidyPerTurn: _outgoingSubsidyPerTurn(game, humanPlayerId, id),
-        pendingGrantAmount: econ.grant,
-        pendingSubsidyAmount: econ.subsidy,
-      ),
-    );
-  }
-  minorIds.sort((a, b) => (displayNameFor(a)).compareTo(displayNameFor(b)));
-  for (final id in minorIds) {
-    final econ = _pendingEconomicAmounts(pendingList, id);
-    rows.add(
-      DiplomacyRowData(
-        factionId: id,
-        displayName: displayNameFor(id),
-        kind: FactionKind.minor,
-        relation: view.diplomacyByOtherId[id],
-        overture: getOverture(game, humanPlayerId, id),
-        actions: actionsByTarget[id] ?? [],
-        powerScore: null,
-        playerPowerScore: null,
-        pendingOrderTypes: pendingByTarget[id] ?? {},
-        activeSubsidyPerTurn: _outgoingSubsidyPerTurn(game, humanPlayerId, id),
-        pendingGrantAmount: econ.grant,
-        pendingSubsidyAmount: econ.subsidy,
-      ),
-    );
-  }
-  tribeIds.sort((a, b) => (displayNameFor(a)).compareTo(displayNameFor(b)));
-  for (final id in tribeIds) {
-    final econ = _pendingEconomicAmounts(pendingList, id);
-    rows.add(
-      DiplomacyRowData(
-        factionId: id,
-        displayName: displayNameFor(id),
-        kind: FactionKind.tribe,
-        relation: view.diplomacyByOtherId[id],
-        overture: getOverture(game, humanPlayerId, id),
-        actions: actionsByTarget[id] ?? [],
-        powerScore: null,
-        playerPowerScore: null,
-        pendingOrderTypes: pendingByTarget[id] ?? {},
-        activeSubsidyPerTurn: _outgoingSubsidyPerTurn(game, humanPlayerId, id),
-        pendingGrantAmount: econ.grant,
-        pendingSubsidyAmount: econ.subsidy,
-      ),
-    );
-  }
-  return rows;
-}
+export 'diplomacy_panel_rows.dart';
 
 /// Full-page diplomacy panel. SPEC/ui/diplomacy-panel.md.
 class DiplomacyPanel extends StatefulWidget {
